+++ FoliCon/Modules/DialogServiceExtensions.cs
@@ -1,4 +1,5 @@
-﻿using NLog;
+﻿using FoliCon.Modules.TMDB;
+using NLog;
 using Logger = NLog.Logger;
 
 namespace FoliCon.Modules;
