+++ FoliCon/ViewModels/PosterPickerViewModel.cs
@@ -1,4 +1,5 @@
-﻿using FoliCon.Modules.utils;
+﻿using FoliCon.Modules.TMDB;
+using FoliCon.Modules.utils;
 using NLog;
 using Collection = TMDbLib.Objects.Collections.Collection;
 using Logger = NLog.Logger;
@@ -100,7 +101,7 @@ public class PosterPickerViewModel : BindableBase, IDialogAware
         LoadData();
             
     }
-    public void LoadData()
+    public async void LoadData()
     {
         var resultType = Result.MediaType;
         var response = _isPickedById
@@ -114,19 +115,19 @@ public class PosterPickerViewModel : BindableBase, IDialogAware
             {
                 dynamic pickedResult = _isPickedById ? (TvShow)response : (SearchTv)response;
                 Title = pickedResult.Name;
-                images = TmdbObject.SearchTvImages(pickedResult.Id);
+                images = await TmdbObject.SearchTvImages(pickedResult.Id);
             }
             else if (resultType == MediaTypes.Movie)
             {
                 dynamic pickedResult = _isPickedById ? (Movie)response : (SearchMovie)response;
                 Title = pickedResult.Title;
-                images = TmdbObject.SearchMovieImages(pickedResult.Id);
+                images = await TmdbObject.SearchMovieImages(pickedResult.Id);
             }
             else if (resultType == MediaTypes.Collection)
             {
                 dynamic pickedResult = _isPickedById ? (Collection)response : (SearchCollection)response;
                 Title = pickedResult.Name;
-                images = TmdbObject.SearchCollectionImages(pickedResult.Id);
+                images = await TmdbObject.SearchCollectionImages(pickedResult.Id);
             }
             else if (resultType == MediaTypes.Mtv)
             {
@@ -137,14 +138,14 @@ public class PosterPickerViewModel : BindableBase, IDialogAware
                     {
                         SearchTv pickedResult = response;
                         Title = pickedResult.Name;
-                        images = TmdbObject.SearchTvImages(pickedResult.Id);
+                        images = await TmdbObject.SearchTvImages(pickedResult.Id);
                         break;
                     }
                     case MediaType.Movie:
                     {
                         SearchMovie pickedResult = response;
                         Title = pickedResult.Title;
-                        images = TmdbObject.SearchMovieImages(pickedResult.Id);
+                        images = await TmdbObject.SearchMovieImages(pickedResult.Id);
                         break;
                     }
                 }
